test: ensure worker picks up requests once up
If the worker is dead and the request queue gets new rows,
these will get processed once the worker is up again.
