+++ test/init.sql
@@ -1,4 +1,7 @@
 create database pre_existing;
 create role pre_existing nosuperuser login;
+
+\c postgres
 create extension pg_net;
-\ir ./loadtest/loadtest.sql
+\ir ./utils/loadtest.sql
+\ir ./utils/helpers.sql
